next auth

+++ client/pages/api/deleteArticle.jsx
@@ -1,5 +1,7 @@
 import Cors from "cors"
 import { PrismaClient } from "@prisma/client"
+import { getServerSession } from "next-auth/next"
+import { authOptions } from "./auth/[...nextauth]"
 
 const prisma = new PrismaClient()
 
@@ -21,12 +23,14 @@ function runMiddleware(req, res, fn) {
 
 export default async function deleteArticle(req, res) {
   await runMiddleware(req, res, cors)
+  const session = await getServerSession(req, res, authOptions)
+  if (session) {
+    const { articleId } = req.body
 
-  const { articleId } = req.body
+    const result = await prisma.article.delete({
+      where: { id: articleId },
+    })
 
-  const result = await prisma.article.delete({
-    where: { id: articleId },
-  })
-
-  res.status(200).json(result)
+    res.status(200).json(result)
+  }
 }
